code/requirements.txt
tqdm
pandas
Mastodon.py==1.7.0
tweepy==4.14.0
